Multiupload progressbar

+++ resources/assets/js/components/UploadComponent.vue
@@ -9,10 +9,10 @@
 
                         <div class="row justify-content-center">
                             <div class="upload text-center" v-bind:class="{'upload-dragover' : isDragover}"
-                                 v-on:dragover="ondragover"
-                                 v-on:dragleave="ondragleave"
-                                 v-on:drop="ondrop"
-                                 v-on:click="onclick">
+                                 @dragover="ondragover"
+                                 @dragleave="ondragleave"
+                                 @drop="ondrop"
+                                 @click="onclick">
 
                                 <i class="fas fa-cloud-upload-alt fa-5x text-secondary"></i>
                                 <p class="text-secondary">Drop files here or click to choose files...</p>
@@ -21,45 +21,8 @@
                                 </div>
                             </div>
                         </div>
-                        <div class="row align-content-center text-center gallery">
-                            <div class="col-md-4" v-for="media in medias">
 
-                                <div class="img-thumb-error" v-if="!media.isImage">
-                                    <div class="overlay"></div>
-                                    <i class="fas fa-exclamation-triangle text-danger fa-7x pt-4"></i>
-                                    <p class="text-danger font-weight-bold">File type not supported. - {{media.name}}</p>
-                                    <div class="img-thumb-button">
-                                        <button type="button" class="btn btn-danger" @click="deleteMedia(media.id)">
-                                            <i class="fas fa-trash-alt"></i>
-                                        </button>
-                                    </div>
-                                </div>
-                                <div v-else>
-                                    <div class="img-thumb-error" v-if="media.isOverMaxSize">
-                                        <div class="overlay"></div>
-                                        <i class="fas fa-exclamation-triangle text-danger fa-7x pt-4"></i>
-                                        <p class="text-danger font-weight-bold">File size exceeded. - {{media.name}}</p>
-                                        <div class="img-thumb-button">
-                                            <button type="button" class="btn btn-danger" @click="deleteMedia(media.id)">
-                                                <i class="fas fa-trash-alt"></i>
-                                            </button>
-                                        </div>
-                                    </div>
-                                    <div class="img-thumb" v-else>
-                                        <div class="overlay"></div>
-                                        <img :src="media.media">
-                                        <div class="img-thumb-button">
-                                            <button type="button" class="btn btn-primary" @click="showPopup(media.media)">
-                                                <i class="fas fa-search"></i>
-                                            </button>
-                                            <button type="button" class="btn btn-danger" @click="deleteMedia(media.id)">
-                                                <i class="fas fa-trash-alt"></i>
-                                            </button>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
+                        <gallery-list-component :medias="medias" v-on:delete-media="deleteMedia"></gallery-list-component>
                     </div>
                 </div>
             </div>
@@ -72,6 +35,8 @@
 </template>
 
 <script>
+    import GalleryListComponent from './GalleryComponent';
+
     export default {
         data : function(){
             return {
@@ -79,9 +44,15 @@
                 medias : [],
             }
         },
+
+        components : {
+            GalleryListComponent
+        },
+
         created(){
             this.fetchData();
         },
+
         methods : {
             ondragover : function(e) {
                 e.preventDefault();
@@ -103,26 +74,44 @@
                 this.$refs.files.click();
             },
             onchange : function(e){
-              const files = e.target.files;
-              this.handleUpload(files);
+                const files = e.target.files;
+                this.handleUpload(files);
             },
             handleUpload : function (files){
-                let formData = new FormData();
-                const config = { onUploadProgress: progressEvent => {
-                        console.log(progressEvent.loaded / progressEvent.total * 100);
-                    }
-                };
 
                 Array.from(files).forEach(function(file){
-                    formData.append('files[]', file);
+                    let fileInfo = {
+                        'id' : '',
+                        'name' : file.name,
+                        'media' : '',
+                        'isImage' : file.type == 'image/jpeg' || file.type == 'image/png' ? true : false,
+                        'isOverMaxSize' : file.size > 10485760 ? true : false,
+                        'isUpload' : true,
+                        'progessPercentage' : 0,
+                    };
 
-                });
+                    this.medias.push(fileInfo);
+                    if( fileInfo.isImage && !fileInfo.isOverMaxSize){
+                        const formData = new FormData();
+                        formData.append('files[]', file);
 
-                axios.post('media-upload', formData, config)
-                    .then(response => {
-                        this.$refs.files = "";
-                        this.medias = this.medias.concat(response.data.data);
-                    });
+                        axios.post('media-upload', formData, { onUploadProgress: progressEvent => {
+                                fileInfo.progessPercentage = parseInt( Math.round( ( progressEvent.loaded / progressEvent.total * 100)));
+                            }
+                        })
+                            .then(response => {
+                                fileInfo.id = response.data.data.id;
+                                fileInfo.name = response.data.data.name;
+                                fileInfo.media = response.data.data.media;
+                                fileInfo.isImage = response.data.data.isImage;
+                                fileInfo.isOverMaxSize = response.data.data.isOverMaxSize;
+                                fileInfo.isUpload = response.data.data.isUpload;
+                                fileInfo.progessPercentage = response.data.data.progessPercentage;
+                            });
+                    }
+                }.bind(this));
+
+                this.$refs.files.value = '';
 
             },
             fetchData : function(){
@@ -131,15 +120,14 @@
                         this.medias = response.data.data;
                     });
             },
-            showPopup: function(url){
-                $('#myModal').modal('show');
-                $('#img-popup').attr('src',url);
-            },
-            deleteMedia : function(id){
-                axios.delete('media-upload/'+id)
-                    .then(response => {
-                        this.fetchData();
-                    })
+            deleteMedia : function(id, index){
+                this.medias.splice(index, 1);
+                if(id != ''){
+                    axios.delete('media-upload/'+id)
+                        .then(response => {
+                            console.log(response.data.msg);
+                        });
+                }
             }
 
         }
@@ -163,56 +151,8 @@
         border: 2px dashed #4e555b;
     }
 
-    .gallery{
-        padding-top: 20px;
-    }
-
     .wrapper{
         display: none;
-    }
-
-    .img-thumb{
-        width: 200px;
-        height: 200px;
-
-        img{
-            width: 100%;
-            height: auto;
-        }
-    }
-
-    .img-thumb-error{
-        width: 200px;
-        height: 200px;
-    }
-
-    .overlay{
-        position: absolute;
-        width: 200px;
-        height: 200px;
-        padding: 10px 10px;
-
-    }
-
-    .img-thumb:hover .overlay,
-    .img-thumb-error:hover .overlay{
-        display: block;
-        background: #1d2124;
-        opacity: 0.5;
-    }
-
-    .img-thumb-button {
-        opacity: 0;
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        text-align: center;
-    }
-
-    .img-thumb:hover .img-thumb-button,
-    .img-thumb-error:hover .img-thumb-button{
-        opacity: 1;
     }
 
     .modal-dialog {
@@ -239,4 +179,5 @@
         text-decoration: none;
         cursor: pointer;
     }
+
 </style>
